+++ cypress/fixtures/outline.feature.ts
@@ -13,6 +13,11 @@ Then('I should have {int} cucumbers', (a: number) => {
 });
 
 feature('Examples Tables in Gherkin text', () => {
+  // Sometimes it can be desireable to run the same scenario multiple times
+  // with different data each time. This can be done by placing an Examples
+  // section with an Examples Table underneath a Scenario, and use <placeholders>
+  // in the Scenario, matching the table headers.
+
   scenarioOutline('eating cucumbers', () => {
     outline(() => {
       given('there are <start> cucumbers');