+++ cypress/fixtures/rules.feature.ts
@@ -25,14 +25,22 @@ Given('there are no chocolates inside', () => {
 });
 
 feature('Rules', () => {
+  // You can place scenarios inside rules. This makes is possible to structure
+  // Gherkin documents in the same way as [example maps](https://cucumber.io/blog/bdd/example-mapping-introduction/).
+  // You can also use the Examples synonym for Scenario to make them even more similar.
+
   rule('a sale cannot happen if change cannot be returned', () => {
     example('no change', () => {
+      // Sad path
+
       given('there are 5 0.20 coins inside');
       when('the customer tries to buy a 0.85 chocolate with a 1 coin');
       then('the sale should not happen');
     });
     
     example('exact change', () => {
+      // Happy path
+
       given('there are 5 0.20 coins inside');
       and('there are 3 chocolates inside');
       when('the customer tries to buy a 0.80 chocolate with a 1 coin');
@@ -41,6 +49,8 @@ feature('Rules', () => {
   });
   
   rule('a sale cannot happen if we"re out of stock', () => {
+    // Sad path
+    
     example('no chocolates left', () => {
       given('there are no chocolates inside');
       but('there are 10 0.5 coins inside');